+++ src/signer/default.ts
@@ -7,7 +7,7 @@ import {
   calculateDeployAccountTransactionHash,
   calculateTransactionHash,
 } from '../utils/hash';
-import { toHex } from '../utils/number';
+import { toHex } from '../utils/num';
 import { fromCallsToExecuteCalldata } from '../utils/transaction';
 import { TypedData, getMessageHash } from '../utils/typedData';
 import { SignerInterface } from './interface';