+++ src/types/api/sequencer.ts
@@ -1,5 +1,5 @@
 import { BlockIdentifier } from '../../provider/utils';
-import { BigNumberish } from '../../utils/number';
+import { BigNumberish } from '../../utils/num';
 import {
   Abi,
   AllowArray,
@@ -251,7 +251,7 @@ export namespace Sequencer {
     fee_estimation: Sequencer.EstimateFeeResponse;
   };
 
-  export type SimulateTransaction = Omit<InvokeFunctionTransaction, 'max_fee' | 'entry_point_type'>;
+  export type SimulateTransaction = Omit<InvokeFunctionTransaction, 'entry_point_type'>;
 
   export type EstimateFeeRequestBulk = AllowArray<
     InvokeEstimateFee | DeclareEstimateFee | DeployEstimateFee | DeployAccountEstimateFee