+++ src/types/lib.ts
@@ -1,5 +1,5 @@
 import { weierstrass } from '../utils/ec';
-import type { BigNumberish } from '../utils/number';
+import type { BigNumberish } from '../utils/num';
 
 // Common Signature Type which needs to be imported from weierstrass
 // and imported at many places
@@ -177,15 +177,29 @@ export type Abi = Array<FunctionAbi | EventAbi | StructAbi>;
 
 type EventAbi = any;
 
+export type Builtins = string[];
+
+export type SieraContractEntryPointFields = {
+  selector: string;
+  function_idx: number;
+};
+
 export type ContractEntryPointFields = {
   selector: string;
   offset: string;
+  builtins?: Builtins;
 };
 
 export type EntryPointsByType = {
   CONSTRUCTOR: ContractEntryPointFields[];
   EXTERNAL: ContractEntryPointFields[];
   L1_HANDLER: ContractEntryPointFields[];
+};
+
+export type SieraEntryPointsByType = {
+  CONSTRUCTOR: SieraContractEntryPointFields[];
+  EXTERNAL: SieraContractEntryPointFields[];
+  L1_HANDLER: SieraContractEntryPointFields[];
 };
 
 export interface Program extends Record<string, any> {
@@ -200,6 +214,24 @@ export type CompiledContract = LegacyCompiledContract | SieraContractClass;
 
 export type LegacyCompiledContract = Omit<LegacyContractClass, 'program'> & {
   program: Program;
+};
+
+export type Hints = [number, string[]][];
+
+export type CompiledSieraCasm = {
+  prime: string;
+  compiler_version: string;
+  bytecode: string[];
+  hints: Hints;
+  entry_points_by_type: EntryPointsByType;
+};
+
+export type CompiledSiera = {
+  sierra_program: string[];
+  sierra_program_debug_info: any;
+  contract_class_version: string;
+  entry_points_by_type: SieraEntryPointsByType;
+  abi: Abi;
 };
 
 export type Struct = {